Renamed CheckNewMessagesSyncTask to CheckNewMessagesInInboxWorker and moved to jetpack.workmanager.sync.| #993
