+++ FlowCrypt/src/main/java/com/flowcrypt/email/service/IdleService.kt
@@ -15,11 +15,11 @@ import androidx.lifecycle.liveData
 import androidx.lifecycle.switchMap
 import com.flowcrypt.email.api.email.model.LocalFolder
 import com.flowcrypt.email.api.email.sync.IdleSyncRunnable
-import com.flowcrypt.email.api.email.sync.tasks.CheckNewMessagesSyncTask
 import com.flowcrypt.email.database.FlowCryptRoomDatabase
 import com.flowcrypt.email.database.entity.AccountEntity
 import com.flowcrypt.email.jetpack.lifecycle.ConnectionLifecycleObserver
 import com.flowcrypt.email.jetpack.viewmodel.AccountViewModel
+import com.flowcrypt.email.jetpack.workmanager.sync.CheckNewMessagesInInboxWorker
 import com.flowcrypt.email.util.LogsUtil
 import com.flowcrypt.email.util.exception.ExceptionUtil
 import com.sun.mail.imap.IMAPFolder
@@ -125,7 +125,7 @@ class IdleService : LifecycleService() {
 
       override fun messagesAdded(accountEntity: AccountEntity, localFolder: LocalFolder, e: MessageCountEvent?) {
         LogsUtil.d(TAG, "messagesAdded: " + e?.messages?.size)
-        CheckNewMessagesSyncTask.enqueue(applicationContext)
+        CheckNewMessagesInInboxWorker.enqueue(applicationContext)
       }
 
       override fun messagesRemoved(accountEntity: AccountEntity, localFolder: LocalFolder, e: MessageCountEvent?) {
